% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwanda_theme.R
\name{fancy_scientific}
\alias{fancy_scientific}
\title{this function is taken from this thread:
https://groups.google.com/forum/#!topic/ggplot2/a_xhMoQyxZ4}
\usage{
fancy_scientific(l)
}
\description{
it puts axis labels in scientific notation
}

