% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwtheme_package.R
\docType{package}
\name{rwtheme}
\alias{package-rwtheme}
\alias{rwtheme}
\alias{rwtheme-package}
\title{ggplot2 theme for Rwanda survey}
\description{
This is a \pkg{ggplot2} theme that has been customized for the Rwanda
2011 survey. It is inspired by the \pkg{ggthemes} package, which has
several nice themes collected together in package format.
}

