% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwanda_theme.R
\name{scale_sex_values}
\alias{scale_color_sex}
\alias{scale_fill_sex}
\alias{scale_sex_values}
\title{sex color scale}
\usage{
scale_sex_values()

scale_color_sex(...)

scale_fill_sex(...)
}
\description{
sex color scale

sex color scale

sex color scale
}

