% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwanda_theme.R
\name{scale_tiedefn_values}
\alias{scale_color_tiedefn}
\alias{scale_fill_tiedefn}
\alias{scale_linetype_tiedefn}
\alias{scale_tiedefn_values}
\title{tie definition color scale}
\usage{
scale_tiedefn_values()

scale_color_tiedefn(...)

scale_linetype_tiedefn(...)

scale_fill_tiedefn(...)
}
\description{
tie definition color scale

tie definition color scale

tie definition linetype scale

tie definition color scale
}

