% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwanda_theme.R
\name{theme_rwanda}
\alias{theme_rwanda}
\title{theme for Rwanda plots}
\usage{
theme_rwanda(base_size = 12, base_family = "")
}
\description{
see the source for ggplot2::theme_bw() to see how we're doing this
}

